Mostly working, I think something might be wonky about either the way the distance from centroid to pixel is calculated, or the way the centroids are displayed. The colors of the centroids seem right, but sometimes they seem too close to eachother.
Need to cast out lines from each centroid to the pixels it owns, and also make
the radius of the centroid the distance from the centroid to it's most distant pixel.

As i write this it occurs to me that the position of a sphere might be the upper right corner of the bounding box
instead of the center, need to read up on the threejs docs to find out.